Ny struktur for kontraktrepoet. - modellert etter familie-felles - 4 moduler: felles, kontantstotte, barnetrygd, enslig forsørger
