+++ felles/src/main/kotlin/no/nav/familie/kontrakter/felles/Kontrakt.kt
@@ -1,0 +1,23 @@
+package no.nav.familie.kontrakter.felles
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect
+import com.fasterxml.jackson.annotation.PropertyAccessor
+import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.SerializationFeature
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+
+interface Kontrakt
+
+val objectMapper: ObjectMapper
+    get() = ObjectMapper()
+        .setVisibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.NONE)
+        .setVisibility(PropertyAccessor.SETTER, JsonAutoDetect.Visibility.NONE)
+        .setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY)
+        .setVisibility(PropertyAccessor.CREATOR, JsonAutoDetect.Visibility.ANY)
+        .registerKotlinModule()
+        .registerModule(JavaTimeModule())
+        .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+        .disable(SerializationFeature.FAIL_ON_EMPTY_BEANS)
+        .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
