+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/dokarkiv/api/ArkiverDokumentResponse.kt
@@ -1,0 +1,14 @@
+package no.nav.familie.kontrakter.ks.dokarkiv.api
+
+import com.fasterxml.jackson.module.kotlin.readValue
+import no.nav.familie.kontrakter.felles.Kontrakt
+import no.nav.familie.kontrakter.felles.objectMapper
+
+class ArkiverDokumentResponse(val journalpostId: String, private val ferdigstilt: Boolean?) : Kontrakt {
+    val isFerdigstilt: Boolean
+        get() = ferdigstilt!!
+}
+
+fun String.toArkiverDokumentResponse(): ArkiverDokumentResponse = objectMapper.readValue(this)
+
+fun ArkiverDokumentResponse.toJson(): String = objectMapper.writeValueAsString(this)
