+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/dokarkiv/api/DokumentType.kt
@@ -1,0 +1,6 @@
+package no.nav.familie.kontrakter.ks.dokarkiv.api
+
+enum class DokumentType {
+    KONTANTSTØTTE_SØKNAD,
+    KONTANTSTØTTE_SØKNAD_VEDLEGG
+}
