+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/dokarkiv/api/FilType.kt
@@ -1,0 +1,6 @@
+package no.nav.familie.kontrakter.ks.dokarkiv.api
+
+enum class FilType {
+    PDFA,
+    JSON
+}
