+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/søknad/OppgittErklæring.kt
@@ -1,0 +1,8 @@
+package no.nav.familie.kontrakter.ks.søknad
+
+data class OppgittErklæring(
+        val isBarnetHjemmeværendeOgIkkeAdoptert: Boolean,
+        val isBorSammenMedBarnet: Boolean,
+        val isIkkeAvtaltDeltBosted: Boolean,
+        val isBarnINorgeNeste12Måneder: Boolean
+)
