+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/søknad/OppgittFamilieforhold.kt
@@ -1,0 +1,7 @@
+package no.nav.familie.kontrakter.ks.søknad
+
+data class OppgittFamilieforhold(
+    val barna: Set<Barn>,
+    val borBeggeForeldreSammen: Boolean,
+    val oppgittAnnenPartNavn: String?
+)
