+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/søknad/OppgittUtlandsTilknytning.kt
@@ -1,0 +1,6 @@
+package no.nav.familie.kontrakter.ks.søknad
+
+data class OppgittUtlandsTilknytning(
+        val aktørerArbeidYtelseIUtlandet: Set<AktørArbeidYtelseUtland>,
+        val aktørerTilknytningTilUtlandet: Set<AktørTilknytningUtland>
+)
