+++ kontantstotte/src/main/kotlin/no/nav/familie/kontrakter/ks/søknad/Standpunkt.kt
@@ -1,0 +1,5 @@
+package no.nav.familie.kontrakter.ks.søknad
+
+enum class Standpunkt {
+    JA, NEI, UBESVART
+}
