+++ pom.xml
@@ -4,12 +4,12 @@
 
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>no.nav.familie.ks</groupId>
-    <artifactId>kontrakt</artifactId>
+    <groupId>no.nav.familie</groupId>
+    <artifactId>kontrakter</artifactId>
     <version>${revision}${sha1}${changelist}</version>
-    <packaging>jar</packaging>
-    <description>Kontantstøtte DTO'er</description>
-    <url>https://github.com/navikt/familie-ks-kontrakt</url>
+    <packaging>pom</packaging>
+    <description>Kontrakter for familieytelsene</description>
+    <url>https://github.com/navikt/familie-kontrakter</url>
 
     <name>KS Kontrakt</name>
 
@@ -18,11 +18,18 @@
         <sha1></sha1>
         <changelist>-SNAPSHOT</changelist>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <kotlin.version>1.3.50</kotlin.version>
+        <kotlin.version>1.3.61</kotlin.version>
         <jackson.version>2.9.10</jackson.version>
         <kotlin.code.style>official</kotlin.code.style>
         <junit.version>4.12</junit.version>
     </properties>
+
+    <modules>
+        <module>felles</module>
+        <module>kontantstotte</module>
+        <!--module>ba</module-->
+        <!--module>ef</module-->
+    </modules>
 
     <dependencies>
         <dependency>
@@ -88,59 +95,6 @@
         </dependency>
     </dependencies>
 
-    <build>
-        <sourceDirectory>src/main/kotlin</sourceDirectory>
-        <testSourceDirectory>src/test/kotlin</testSourceDirectory>
-
-        <plugins>
-            <plugin>
-                <groupId>org.jetbrains.kotlin</groupId>
-                <artifactId>kotlin-maven-plugin</artifactId>
-                <version>${kotlin.version}</version>
-                <executions>
-                    <execution>
-                        <id>compile</id>
-                        <phase>compile</phase>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                    </execution>
-                    <execution>
-                        <id>test-compile</id>
-                        <phase>test-compile</phase>
-                        <goals>
-                            <goal>test-compile</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <version>3.1.1</version>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.0</version>
-                <configuration>
-                    <release>11</release>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-release-plugin</artifactId>
-                <version>2.5.3</version>
-                <configuration>
-                    <autoVersionSubmodules>true</autoVersionSubmodules>
-                    <useReleaseProfile>true</useReleaseProfile>
-                    <releaseProfiles>release</releaseProfiles>
-                    <goals>deploy</goals>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
     <distributionManagement>
         <repository>
             <id>github</id>
@@ -149,7 +103,7 @@
         </repository>
     </distributionManagement>
 
-    <profiles>
+    <!--profiles>
         <profile>
             <id>release</id>
             <activation>
@@ -159,7 +113,6 @@
             </activation>
             <build>
                 <plugins>
-                    <!--create source jar-->
                     <plugin>
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-source-plugin</artifactId>
@@ -173,8 +126,6 @@
                             </execution>
                         </executions>
                     </plugin>
-
-                    <!--create empty javadoc jar-->
                     <plugin>
                         <groupId>org.apache.maven.plugins</groupId>
                         <artifactId>maven-jar-plugin</artifactId>
@@ -192,11 +143,10 @@
                             </execution>
                         </executions>
                     </plugin>
-
                 </plugins>
             </build>
         </profile>
-    </profiles>
+    </profiles-->
 
     <organization>
         <name>no.nav</name>
