--- src/main/kotlin/no/nav/familie/ks/kontrakter/dokarkiv/api/DokumentType.kt
@@ -1,6 +1,0 @@
-package no.nav.familie.ks.kontrakter.dokarkiv.api
-
-enum class DokumentType {
-    KONTANTSTØTTE_SØKNAD,
-    KONTANTSTØTTE_SØKNAD_VEDLEGG
-}
