--- src/main/kotlin/no/nav/familie/ks/kontrakter/dokarkiv/api/FilType.kt
@@ -1,6 +1,0 @@
-package no.nav.familie.ks.kontrakter.dokarkiv.api
-
-enum class FilType {
-    PDFA,
-    JSON
-}
