--- src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/OppgittFamilieforhold.kt
@@ -1,7 +1,0 @@
-package no.nav.familie.ks.kontrakter.søknad
-
-data class OppgittFamilieforhold(
-    val barna: Set<Barn>,
-    val borBeggeForeldreSammen: Boolean,
-    val oppgittAnnenPartNavn: String?
-)
