--- src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/OppgittUtlandsTilknytning.kt
@@ -1,6 +1,0 @@
-package no.nav.familie.ks.kontrakter.søknad
-
-data class OppgittUtlandsTilknytning(
-        val aktørerArbeidYtelseIUtlandet: Set<AktørArbeidYtelseUtland>,
-        val aktørerTilknytningTilUtlandet: Set<AktørTilknytningUtland>
-)
