--- src/main/kotlin/no/nav/familie/ks/kontrakter/søknad/Standpunkt.kt
@@ -1,5 +1,0 @@
-package no.nav.familie.ks.kontrakter.søknad
-
-enum class Standpunkt {
-    JA, NEI, UBESVART
-}
